Add API to handle deleting recordings  (#21520)
* Add recording delete API

* Re-organize recordings apis

* Fix import

* Consolidate query types
